+++ internal/dto/document_client/get.go
@@ -1,12 +1,12 @@
 package dto
 
 type GetDocumentClientReq struct {
-	Id int64 `json:"id" validate:"required"`
+	Id int32 `json:"id" validate:"required"`
 }
 
 type GetDocumentClientRes struct {
-	Id         int64  `json:"id"`
-	ClientId   int64  `json:"clientId"`
-	DocumentId int64  `json:"documentId"`
+	Id         int32  `json:"id"`
+	ClientId   int32  `json:"clientId"`
+	DocumentId int32  `json:"documentId"`
 	Action     string `json:"action"`
 }
